all done

+++ src/main/java/edu/spps/model/PerformanceModel.java
@@ -1,7 +1,6 @@
 package edu.spps.model;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,13 +8,13 @@ import lombok.Setter;
 @Setter
 @Getter
 public class PerformanceModel {
-  private int id;
-  private int student_id;
-  private String name;  
-  private Integer attendance;
-  private Integer study_hours;
-  private Integer assessment;
-  private Integer participation;
-  private Double percentage;
-  private LocalDate performance_date;
+	private int id;
+	private int student_id;
+	private String name;
+	private double attendance;
+	private double study_hours;
+	private double assessment;
+	private double participation;
+	private Double percentage;
+	private LocalDate performance_date;
 }
